+++ main/i18ntk-init.js
@@ -138,13 +138,15 @@ class I18nInitializer {
       };
       
       const i18nFrameworks = [
-        'react-i18next',
-        'vue-i18n', 
-        'angular-i18n',
-        'i18next',
-        'next-i18next',
-        'svelte-i18n',
-        '@nuxtjs/i18n'
+        'react-i18next', 'vue-i18n', 'i18next', 'next-i18next',
+        'svelte-i18n', '@nuxtjs/i18n', 'next-intl', 'remix-i18next',
+        'i18next-remix', 'gatsby-plugin-react-i18next', 'gatsby-plugin-intl',
+        'sveltekit-i18n', 'astro-i18next', '@astrojs/i18n',
+        'qwik-speak', 'qwik-i18n', '@solid-primitives/i18n',
+        'ember-intl', 'react-native-localize', 'expo-localization',
+        '@ngx-translate/core', '@ionic/angular',
+        'formatjs', '@lingui/core', 'i18ntk-runtime', 'i18ntk/runtime',
+        'angular-i18n', '@angular/localize'
       ];
       
       const installedFrameworks = i18nFrameworks.filter(framework => dependencies[framework]);